+++ src/icepool/__init__.py
@@ -17,7 +17,7 @@ General conventions:
 
 __docformat__ = 'google'
 
-__version__ = '1.5.1a1'
+__version__ = '1.6.0a0'
 
 from typing import Final
 
@@ -48,8 +48,8 @@ Operation in specific cases:
 
 from icepool.function import (d, z, __getattr__, coin, stochastic_round,
                               one_hot, iter_cartesian_product, from_cumulative,
-                              from_rv, min_outcome, max_outcome, sorted_union, align,
-                              align_range, commonize_denominator, reduce,
+                              from_rv, min_outcome, max_outcome, consecutive,
+                              sorted_union, commonize_denominator, reduce,
                               accumulate, map, map_function, map_and_time,
                               map_to_pool)
 
@@ -147,10 +147,10 @@ __all__ = [
     'd', 'z', 'coin', 'stochastic_round', 'one_hot', 'Outcome', 'Die',
     'Population', 'tupleize', 'vectorize', 'Vector', 'Symbols', 'Again',
     'CountsKeysView', 'CountsValuesView', 'CountsItemsView', 'from_cumulative',
-    'from_rv', 'lowest', 'highest', 'middle', 'min_outcome', 'max_outcome', 'sorted_union',
-    'align', 'align_range', 'commonize_denominator', 'reduce', 'accumulate',
-    'map', 'map_function', 'map_and_time', 'map_to_pool', 'Reroll',
-    'RerollType', 'Pool', 'standard_pool', 'MultisetGenerator', 'Alignment',
+    'from_rv', 'lowest', 'highest', 'middle', 'min_outcome', 'max_outcome',
+    'consecutive', 'sorted_union', 'commonize_denominator', 'reduce',
+    'accumulate', 'map', 'map_function', 'map_and_time', 'map_to_pool',
+    'Reroll', 'RerollType', 'Pool', 'standard_pool', 'MultisetGenerator',
     'MultisetExpression', 'MultisetEvaluator', 'Order', 'Deck', 'Deal',
     'MultiDeal', 'multiset_function', 'function', 'typing', 'evaluator'
 ]