+++ src/icepool/evaluator/__init__.py
@@ -7,16 +7,18 @@ from icepool.evaluator.basic import (ExpandEvaluator, SumEvaluator,
                                      count_evaluator, AnyEvaluator,
                                      any_evaluator)
 from icepool.evaluator.poker import (
-    HighestOutcomeAndCountEvaluator, LargestCountEvaluator,
-    LargestCountAndOutcomeEvaluator, CountSubsetEvaluator, AllCountsEvaluator,
-    LargestStraightEvaluator, LargestStraightAndOutcomeEvaluator,
-    AllStraightsEvaluator, AllStraightsReduceCountsEvaluator)
+    HighestOutcomeAndCountEvaluator, highest_outcome_and_count_evaluator,
+    LargestCountEvaluator, largest_count_evaluator,
+    LargestCountAndOutcomeEvaluator, largest_count_and_outcome_evaluator,
+    CountSubsetEvaluator, AllCountsEvaluator, LargestStraightEvaluator,
+    largest_straight_evaluator, LargestStraightAndOutcomeEvaluator,
+    largest_straight_and_outcome_evaluator, AllStraightsEvaluator,
+    all_straights_evaluator, AllStraightsReduceCountsEvaluator)
 from icepool.evaluator.comparison import (
     ComparisonEvaluator, IsSubsetEvaluator, IsProperSubsetEvaluator,
     IsSupersetEvaluator, IsProperSupersetEvaluator, IsEqualSetEvaluator,
     IsNotEqualSetEvaluator, IsDisjointSetEvaluator)
 from icepool.evaluator.joint import JointEvaluator
-from icepool.evaluator.constant import ConstantEvaluator
 from icepool.evaluator.keep import KeepEvaluator
 from icepool.evaluator.argsort import ArgsortEvaluator
 from icepool.evaluator.expression import ExpressionEvaluator
@@ -24,13 +26,15 @@ from icepool.evaluator.expression import ExpressionEvaluator
 __all__ = [
     'JointEvaluator', 'ExpandEvaluator', 'SumEvaluator', 'sum_evaluator',
     'CountEvaluator', 'count_evaluator', 'AnyEvaluator', 'any_evaluator',
-    'HighestOutcomeAndCountEvaluator', 'LargestCountEvaluator',
-    'LargestCountAndOutcomeEvaluator', 'CountSubsetEvaluator',
-    'AllCountsEvaluator', 'LargestStraightEvaluator',
-    'LargestStraightAndOutcomeEvaluator', 'AllStraightsEvaluator',
-    'AllStraightsReduceCountsEvaluator', 'ComparisonEvaluator',
-    'IsSubsetEvaluator', 'IsProperSubsetEvaluator', 'IsSupersetEvaluator',
-    'IsProperSupersetEvaluator', 'IsEqualSetEvaluator',
-    'IsNotEqualSetEvaluator', 'IsDisjointSetEvaluator', 'ConstantEvaluator',
-    'KeepEvaluator', 'ArgsortEvaluator', 'ExpressionEvaluator'
+    'HighestOutcomeAndCountEvaluator', 'highest_outcome_and_count_evaluator',
+    'LargestCountEvaluator', 'largest_count_evaluator',
+    'LargestCountAndOutcomeEvaluator', 'largest_count_and_outcome_evaluator',
+    'CountSubsetEvaluator', 'AllCountsEvaluator', 'LargestStraightEvaluator',
+    'largest_straight_evaluator', 'LargestStraightAndOutcomeEvaluator',
+    'largest_straight_and_outcome_evaluator', 'AllStraightsEvaluator',
+    'all_straights_evaluator', 'AllStraightsReduceCountsEvaluator',
+    'ComparisonEvaluator', 'IsSubsetEvaluator', 'IsProperSubsetEvaluator',
+    'IsSupersetEvaluator', 'IsProperSupersetEvaluator', 'IsEqualSetEvaluator',
+    'IsNotEqualSetEvaluator', 'IsDisjointSetEvaluator', 'KeepEvaluator',
+    'ArgsortEvaluator', 'ExpressionEvaluator'
 ]