+++ src/icepool/evaluator/joint.py
@@ -78,9 +78,9 @@ class JointEvaluator(MultisetEvaluator[T_contra, tuple]):
         """
         return Order.merge(*(inner.order() for inner in self._inners))
 
-    def alignment(self, outcomes) -> Collection[T_contra]:
-        return icepool.sorted_union(*(evaluator.alignment(outcomes)
-                              for evaluator in self._inners))
+    def extra_outcomes(self, outcomes) -> Collection[T_contra]:
+        return icepool.sorted_union(*(evaluator.extra_outcomes(outcomes)
+                                      for evaluator in self._inners))
 
     @cached_property
     def _prefix_generators(self) -> 'tuple[icepool.MultisetGenerator, ...]':