+++ src/icepool/population/die.py
@@ -87,7 +87,8 @@ class Die(Population[T_co]):
         * Use a dict: `Die({1:1, 2:1, 3:1, 4:1, 5:1, 6:1})`
         * Give the faces as a sequence: `Die([1, 2, 3, 4, 5, 6])`
 
-        All quantities must be non-negative, though they can be zero.
+        All quantities must be non-negative. Outcomes with zero quantity will be
+        omitted.
 
         Several methods and functions foward **kwargs to this constructor.
         However, these only affect the construction of the returned or yielded
@@ -481,42 +482,6 @@ class Die(Population[T_co]):
                 data[max_outcome] += quantity
             else:
                 data[outcome] += quantity
-        return icepool.Die(data)
-
-    def set_range(self: 'Die[int]',
-                  min_outcome: int | None = None,
-                  max_outcome: int | None = None) -> 'Die[int]':
-        """DEPRECATED: Sets the outcomes of this `Die` to the given `int` range (inclusive).
-
-        This may remove outcomes (if they are not within the range)
-        and/or add zero-quantity outcomes (if they are in range but not present
-        in this `Die`).
-
-        Args:
-            min_outcome: The min outcome of the result.
-                If omitted, the min outcome of this `Die` will be used.
-            max_outcome: The max outcome of the result.
-                If omitted, the max outcome of this `Die` will be used.
-        """
-        if min_outcome is None:
-            min_outcome = self.min_outcome()
-        if max_outcome is None:
-            max_outcome = self.max_outcome()
-
-        return self.set_outcomes(range(min_outcome, max_outcome + 1))
-
-    def set_outcomes(self, outcomes: Iterable[T_co]) -> 'Die[T_co]':
-        """DEPRECATED: Sets the set of outcomes to the argument.
-
-        This may remove outcomes (if they are not present in the argument)
-        and/or add zero-quantity outcomes (if they are not present in this `Die`).
-        """
-        data = {x: self.quantity(x) for x in outcomes}
-        return icepool.Die(data)
-
-    def trim(self) -> 'Die[T_co]':
-        """DEPRECATED: Removes all zero-quantity outcomes. """
-        data = {k: v for k, v in self.items() if v > 0}
         return icepool.Die(data)
 
     @cached_property